requirements.txt
Django==3.1.7
psycopg2-binary==2.8.6
Pillow==8.2.0
jsonschema==3.2.0
djangorestframework==3.12.4